backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
asyncpg==0.29.0
pgvector==0.2.4
redis==5.0.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
lightfm==1.17
httpx==0.25.2
beautifulsoup4==4.12.2
# Безопасность (Исправленные версии)
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0.post1
celery==5.3.6